Moved O NET classes to GDC instances
Through a python script, I have moved O NET occupation holder classes to instances of a new data item classes. Whether this class will then be moved in a different place of the GDC hierarchy in order to match efforts in OMRSE is still an open discussion. I have also created a new object property to match the hierarchy in O NET (narrower than). The file is in .ttl for readability. I have of course no idea of how to put this into a format that is welcome to the main repo.